+++ src/RadixDlt.NetworkGateway.GatewayApi/Validators/PaginableRequestValidator.cs
@@ -63,11 +63,11 @@
  */
 
 using FluentValidation;
-using RadixDlt.NetworkGateway.GatewayApiSdk.Model;
+using GatewayModel = RadixDlt.NetworkGateway.GatewayApiSdk.Model;
 
 namespace RadixDlt.NetworkGateway.GatewayApi.Validators;
 
-internal class PaginableRequestValidator : AbstractValidator<IPaginableRequest>
+internal class PaginableRequestValidator : AbstractValidator<GatewayModel.IPaginableRequest>
 {
     public PaginableRequestValidator()
     {